+++ src/app/(main)/_components/Banner/BannerRightRealTimeTalk.tsx
@@ -17,7 +17,9 @@ interface ReviewListType {
   ReviewList: BannerReview[];
 }
 
-export default function RealTimeHotTalk({ ReviewList }: ReviewListType) {
+export default function BannerRightRealTimeTalk({
+  ReviewList,
+}: ReviewListType) {
   return (
     <div className="flex  flex-col gap-[20px]  text-Silver Tablet:justify-end">
       <h2 className=" hidden Laptop:block Laptop:w-[390px]  Laptop:Text-xl-Bold Desktop:w-[521px]  ">
