+++ src/app/(main)/_components/Skeleton/Keyword/LaptopKeyword.tsx
