+++ src/app/(main)/_components/Skeleton/Keyword/MobileKeyword.tsx
