+++ src/app/(main)/_components/Skeleton/MasterPieceMovie/LaptopMasterPiece.tsx
