+++ src/app/(main)/_components/Skeleton/MovieTop/MobileMovieTop.tsx
